+++ geniuspace/database/seeders/DatabaseSeeder.php
@@ -17,5 +17,4 @@ class DatabaseSeeder extends Seeder
         $this->call(ShowcaseSeeder::class);
         $this->call(DualWorldsSeeder::class);
     }
-    }
 }